fix: FIX-068 — corrigir minutes_from_quota para armazenar minutos reais (granularidade 30s)
minutes_from_quota armazenava frações de 30s (int) em vez de minutos reais,
fazendo uma ligação de 7:27 debitar 15 "minutos" no plano em vez de 7,5.

- Call.minutesFromQuota: Integer → BigDecimal (scale 1)
- CallRepository: sumQuotaMinutes/ByType retornam BigDecimal
- CallCostingService: remove * 2 em resolveQuota; aritmética em BigDecimal
- AuditService: mesma correção; quota e acumuladores em BigDecimal
- AuditCallLineDTO/AuditSummaryDTO: campos de quota migrados para BigDecimal
- Testes atualizados com valores em minutos reais

+++ backend/src/main/java/com/dionialves/AsteraComm/call/CallRepository.java
@@ -62,7 +62,7 @@ public interface CallRepository extends JpaRepository<Call, Long>,
             "AND call_status = 'PROCESSED' " +
             "AND EXTRACT(MONTH FROM call_date) = :month " +
             "AND EXTRACT(YEAR  FROM call_date) = :year", nativeQuery = true)
-    int sumQuotaMinutes(@Param("circuitNumber") String circuitNumber,
+    BigDecimal sumQuotaMinutes(@Param("circuitNumber") String circuitNumber,
             @Param("month") int month,
             @Param("year") int year);
 
@@ -72,7 +72,7 @@ public interface CallRepository extends JpaRepository<Call, Long>,
             "AND call_status = 'PROCESSED' " +
             "AND EXTRACT(MONTH FROM call_date) = :month " +
             "AND EXTRACT(YEAR  FROM call_date) = :year", nativeQuery = true)
-    int sumQuotaMinutesByType(@Param("circuitNumber") String circuitNumber,
+    BigDecimal sumQuotaMinutesByType(@Param("circuitNumber") String circuitNumber,
             @Param("callType") String callType,
             @Param("month") int month,
             @Param("year") int year);
